split version into our propType def
